+++ honeycomb/config.py
@@ -7,3 +7,21 @@ dtype_map = {
     'timedelta': 'INTERVAL'
     # 'category': None
 }
+
+storage_type_specs = {
+    'csv': {
+        'settings': {
+            'index': False,
+            'header': False
+        },
+        'ddl': """
+               ROW FORMAT DELIMITED
+               FIELDS TERMINATED BY ','
+               LINES TERMINATED BY '\\n'
+               """
+    },
+    'pq': {
+        'settings': {},
+        'ddl': 'STORED AS PARQUET'
+    }
+}