+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Table appending logic
+- Functionality to create a table with both table and column comments
+- Can create tables in Parquet format
+- Can generate filenames for new files being added to S3
+
 ## [0.1.0] - 2020-05-11
 
 ### Added